products shown in home page

+++ lib/views/home/models/product_model.dart
@@ -26,7 +26,7 @@ class ProductModel {
       categoryId: data["category_id"],
       name: data['name'],
       categoryName: data['category_name'],
-      price: data['price'].toDouble(),
+      price: data['price'],
       sellingPrice: data["selling_price"],
       images: List<String>.from(data['image_url'] ?? []),
     );
